fix(trusty-mpm): the compaction reload must gate the record removal, not just the slot
`compact_and_release_slot` removed the record from the store before it checked
whether the reloaded record was still a tombstone. Only the slot release sat
behind that check, so a session reactivated by `mark_reactivated` inside
prune's teardown window correctly kept its number and still had its store
record deleted.

`self.get(id)` then answered `SessionNotFound` for a live, running session,
breaking hook correlation, `tm session status`, and every lookup by id, while
`tm ls` rendered it as a `-- deleted --` phantom.

The removal now sits inside the `Some(record)` arm, under the same lock
acquisition and the same reload that decides the release. A record that is no
longer a tombstone at reload time keeps both its record and its slot. The
function returns whether it compacted, so `prune_managed` omits a still-live
session from its outcome instead of reporting `PruneAction::Removed` for a
record still in the store.

The existing test asserted only the slot and passed while the record was being
deleted; its docstring named the gap and waved it off. Both are corrected —
it now asserts the record survives, is still `Active`, and still renders with
its record rather than as a tombstone row.

Splitting `PruneFilter` / `PruneAction` / `PrunedSession` / `PruneOutcome` into
`prune_types.rs` keeps `prune.rs` under the 500-SLOC cap (505 -> 435). They are
re-exported, so no call site moved.

Closes #5912.
