+++ crates/trusty-mpm/changelog.d/5897-prune-releases-slots.md
@@ -1,3 +1,4 @@
 Fixed
 
 - `tm session prune` now frees the slot number behind every tombstone it compacts, so `NUM` stops climbing. `prune_managed`'s compaction branch removed the record from the store and left the slot registry still holding its number; `numbered_snapshot` walks the slots the registry HOLDS and tombstones any whose record has vanished, so each pruned session kept rendering as a `-- deleted --` row and its number was never handed out again. One machine accumulated 44 such rows with new sessions still being assigned into the 130s. `SlotRegistry::release` had exactly one caller, the retention sweep, which is why only sessions whose worktrees had already left the disk ever got their numbers back. The release is gated on the same `workspace_needs_protection` predicate that sweep uses, so a record whose `.worktrees/<uuid>` directory and `.trusty-mpm-worktree` sentinel are still on disk is compacted but keeps its slot — handing that number to a newly-spawned session whose worktree already exists is worse than a cosmetic tombstone row, and the shared predicate keeps the two paths from drifting apart again. Compaction remains a store operation: it removes no directory, and a dry run releases nothing. Both conditions are read from the record the store holds at the moment of removal, not from the snapshot prune took before its loop: a session revived in place while that prune was still tearing other sessions down keeps its number rather than having it handed to the next spawn.
+- `tm session prune` no longer deletes the store record of a session that was reactivated while the prune was running. The reload that decides whether the slot is safe to free now decides the removal as well; previously it gated only the slot, and the record was removed unconditionally before that check ran. A session revived through `tm session resume` mid-prune correctly kept its number and still lost its record, so `tm session status` and every lookup by id answered `SessionNotFound` for a live, running session and `tm ls` rendered it as a `-- deleted --` phantom. A record that is no longer a tombstone at reload time now keeps both its record and its slot, and prune omits it from the reported outcome rather than claiming it was removed.
