+++ crates/trusty-mpm/src/session_manager/prune_slot_tests.rs
@@ -11,7 +11,8 @@
 //! What: the released case (no workspace on disk → slot gone from the listing
 //! and reusable), the protected case (a session-worktree-shaped workspace with
 //! its ownership sentinel → record compacted, slot still held, directory
-//! untouched), and the dry-run case (nothing released).
+//! untouched), the reactivation race (a session revived inside the window keeps
+//! its record AND its number — #5912), and the dry-run case (nothing released).
 //! Test: this file IS the test module; run with `cargo test -p trusty-mpm`.
 
 use std::path::{Path, PathBuf};
@@ -211,30 +212,37 @@ async fn prune_compaction_keeps_the_slot_when_the_worktree_is_still_on_disk() {
     );
 }
 
-/// A record reactivated after the snapshot keeps its slot (#5897).
+/// A session reactivated after the snapshot keeps its record AND its slot
+/// (#5897, #5912).
 ///
 /// Why: `prune_managed` takes ONE `cached_all()` snapshot before its loop and
 /// then tears tmux sessions and git worktrees down inside it, so a tombstone
 /// reached late is compacted long after it was read. `mark_reactivated` revives a
 /// `Decommissioned` record in place, needs no tmux, and can land anywhere in that
-/// window — so the record whose number is about to be handed back can be a LIVE
-/// session by then. Judging that from the snapshot hands a running session's `NUM`
-/// to the next spawn, which is exactly the silent reuse #3034 exists to prevent,
-/// and it is a hazard this release path introduces: before #5897 prune freed no
-/// slot at all. This is the arm that makes the re-read worth its cost —
-/// `workspace_path`, the guard's other input, has no production writer that can
-/// move it under a terminal record.
+/// window — so the record prune is about to compact can be a LIVE session by then.
+/// Two separate things are then wrong to decide from the snapshot. Handing the
+/// number back is the silent reuse #3034 exists to prevent. Deleting the record is
+/// the worse half (#5912): `self.get(id)` starts answering `SessionNotFound` for a
+/// running session, which breaks hook correlation, `tm session status`, and every
+/// lookup by id, while `tm ls` renders it as a `-- deleted --` phantom. Both must
+/// ride on the same reload, which is why this test asserts both — an earlier
+/// revision asserted only the slot and passed while the record was still being
+/// deleted.
+///
+/// This is also the arm that makes the re-read worth its cost — `workspace_path`,
+/// the guard's other input, has no production writer that can move it under a
+/// terminal record.
 ///
 /// What: seeds a Decommissioned tombstone, keeps the body the snapshot would have
 /// carried, observes its slot, then reactivates through the real
 /// `mark_reactivated` — the same `Active` record with `workspace_path` untouched
 /// that an operator's in-pane revival writes — and drives the compaction with the
-/// now-stale snapshot. Asserts the number stays held. It renders as a
-/// `-- deleted --` row because the compaction still removes the record; that
-/// second half of the race is not what this guard governs.
+/// now-stale snapshot. Asserts the record is still retrievable, the number is
+/// still held, and the listing row still carries its record rather than rendering
+/// as a tombstone.
 /// Test: this function IS the test.
 #[tokio::test]
-async fn prune_keeps_the_slot_of_a_record_reactivated_after_the_snapshot() {
+async fn prune_keeps_the_record_and_slot_of_a_session_reactivated_after_the_snapshot() {
     let dir = crate::test_support::hermetic_temp_dir();
     let (mgr, _fake) = make_manager(&dir).await;
 
@@ -258,9 +266,25 @@ async fn prune_keeps_the_slot_of_a_record_reactivated_after_the_snapshot() {
         .expect("reactivate in place");
 
     let names = super::super::decommission::worktree_dir_names();
-    mgr.compact_and_release_slot(&stale, &names)
+    let compacted = mgr
+        .compact_and_release_slot(&stale, &names)
         .await
         .expect("compaction");
+    assert!(
+        !compacted,
+        "a reactivated session is not compacted, so prune must not report it as removed"
+    );
+
+    // #5912: the record itself, not just its number.
+    let live = mgr
+        .get(&id)
+        .await
+        .expect("a live, just-reactivated session's record must survive compaction");
+    assert_eq!(
+        live.state,
+        ManagedSessionState::Active,
+        "the surviving record is the reactivated body, not a resurrected tombstone"
+    );
 
     let after = mgr.numbered_snapshot(&mgr.list().await).await;
     assert!(
@@ -268,6 +292,15 @@ async fn prune_keeps_the_slot_of_a_record_reactivated_after_the_snapshot() {
         "a reactivated session's number must not be handed back for reuse; \
          still held: {:?}",
         after.iter().map(|s| s.slot).collect::<Vec<_>>()
+    );
+    let row = after
+        .iter()
+        .find(|s| s.slot == slot)
+        .expect("slot still held");
+    assert_eq!(
+        row.record.as_ref().map(|r| r.id),
+        Some(id),
+        "the held slot must render its live record, not a `-- deleted --` phantom"
     );
 }
